#2219 Fix - change filtering data from pandas to polars API

See merge request noc/noc!7521

(cherry picked from commit 8ae828397497b7aa32a984756b197e2d724f08aa)

bdf8767c Fix - change filtering data from pandas to polars API
74438249 Refactor

+++ core/datasources/interfacemacsstatds.py
@@ -32,7 +32,7 @@ class InterfaceMACsStatDS(BaseDataSource):
     name = "interfacemacsstatds"
 
     fields = [
-        FieldInfo(name="managed_object_id", type=FieldType.UINT),
+        FieldInfo(name="managed_object_id", type=FieldType.UINT64),
         FieldInfo(name="interface_name"),
         FieldInfo(name="mac_count", type=FieldType.UINT),
     ]
